group code between node and block controller
