+++ BackofficeBundle/Controller/BlockController.php
@@ -2,19 +2,16 @@
 
 namespace OpenOrchestra\BackofficeBundle\Controller;
 
-use OpenOrchestra\Backoffice\NavigationPanel\Strategies\TreeNodesPanelStrategy;
 use OpenOrchestra\ModelInterface\Event\NodeEvent;
-use OpenOrchestra\ModelInterface\Model\NodeInterface;
 use OpenOrchestra\ModelInterface\NodeEvents;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration as Config;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use OpenOrchestra\Backoffice\NavigationPanel\Strategies\TransverseNodePanelStrategy;
 
 /**
  * Class BlockController
  */
-class BlockController extends AbstractAdminController
+class BlockController extends AbstractEditionRoleController
 {
     /**
      * @param Request $request
@@ -41,9 +38,7 @@ class BlockController extends AbstractAdminController
         );
 
         if ($node) {
-            $editionRole = $node->getNodeType() === NodeInterface::TYPE_ERROR? TreeNodesPanelStrategy::ROLE_ACCESS_UPDATE_ERROR_NODE:TreeNodesPanelStrategy::ROLE_ACCESS_UPDATE_NODE;
-            $editionRole = $node->getNodeType() === NodeInterface::TYPE_TRANSVERSE? TransverseNodePanelStrategy::ROLE_ACCESS_UPDATE_GENERAL_NODE:$editionRole;
-            $options['disabled'] = !$this->get('security.authorization_checker')->isGranted($editionRole, $node);
+            $options['disabled'] = !$this->get('security.authorization_checker')->isGranted($this->getEditionRole($node), $node);
         }
 
         $options['method'] = 'POST';
@@ -72,5 +67,4 @@ class BlockController extends AbstractAdminController
             $form->getConfig()->getAttribute('template')
         );
     }
-
 }
